Added skip/offset to protocol
UML by default reserves memory at the bottom of the mmaped memory
for kernel code and data. This memory is not mapped into the
mmaped virtual memory. As a result, anything there is not visible.

An external helper should take this into account in order to have
correct memory mappings.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>